requirements.txt
fastapi[all]
fastapi-pagination
uvicorn==0.20.0
SQLAlchemy==2.0.38
pydantic[email]
python-dotenv
psycopg2-binary
asyncpg==0.27.0
alembic==1.9.0
pytest==7.2.0
pytest-asyncio==0.20.3
httpx==0.23.3
pytz
passlib
bcrypt==4.0.1
python-jose
requests
apscheduler